requirements.txt
boto3>=1.35.0
google-cloud-storage-transfer>=1.12.0
google-auth>=2.35.0
google-cloud-pubsub>=2.23.0